\name{logit}
\alias{logit}
\alias{invlogit}

\title{logistic link function}
\description{
Provides the standard logistic link and it's inverse.
}
\usage{
logit(p)
invlogit(eta)
}

\arguments{
  \item{p}{a probability vector}
  \item{eta}{A vector of linear predictors}
  
}
\details{
The usual logisit link and it's inverse.
}

\references{McCullagh and Nelder}
\author{David Pleydell}


\keyword{regression }
