\name{mergeLevels}
\alias{mergeLevels}
\alias{mergeLevels_2}
\title{Evaluates a set of multinomial models}
\description{
  This function makes repeated use of multinomin in the nnet library.
  The set of models is defined by a specified factor in the model
  formula and is generated by itteratively merging unique pairs of
  factor levels. In effect multinom is forced to fit the same
  coefficients to the two specified factor levels.

  The motivation is the analysis of rodent trapping data in relation to
  habitat classes. mergeLevels will identify whether or not there is
  redundancy in the habitat classification with respect to trapping
  success. A decrease in AIC (ie a negative delta AIC) following the
  merging of two classes indicates over-parameterisation in the
  original model and redundancy in the classification with respect
  to the available rodent data. 
  
  mergeLevels_2 is an adaptation of mergeLevels as modified by Amelie. I
  didn't yet have time to compare the differences between the two.
}
\usage{
mergeLevels(aFormula, data, mergeFactor, AICtable = FALSE)
mergeLevels_2(aFormula, data, mergeFactor, AICtable = FALSE)
}

\arguments{
  \item{aFormula}{a formula as required by multinom in the nnet library}
  \item{data}{a data frame}
  \item{mergeFactor}{The name of the factor over which merges should be tried.}
  \item{AICtable}{Logical. Should the AIC values be returned in addition
  to the default delta AIC output. FALSE by default.}
}
\details{
The function makes iterative use of multinom. For each itteration factorLevel\_i + factorLevel\_j in the
model formula is replaced with I\(factorLevel\_i + factorLevel\_j\).
  
}
\value{

  \item{delta AIC}{A matrix indicating increase or decrese in AIC in
    relation to the original model. The diagonals are zero because there
    is no merge between level\_i and level\_i.}
  \item{AIC}{Another matrix. The diagonal elements give the AIC off the
    original model.}
}

\author{Amelie Vaniscotte & David Pleydell}

\references{

\href{http://www.ncbi.nlm.nih.gov/pmc/articles/PMC2702787/}{Vaniscotte A., Pleydell D., Raoul F., Quere J.P., Coeurdassier M., Delattre P., Li T., Qian W., Takahashi K., Weidmann J.C., Qiu J., Giraudoux P. 2009 Modelling and spatial discrimination of small mammal assemblages: an example from western Sichuan (China). Ecological Modelling, 220: 1218-1231.}

Raoul F, Pleydell D, Quere JP, Vaniscotte A, Rieffel R, Takahashi K, Bernard N, Wang J, Dobigny T, Galbreath KE, Giraudoux P 2008 Small-mammal assemblage response to deforestation and afforestation in central China. Mammalia 72 : 320-332
}

\note{
  Currently this function does not operate at the species level as
  multignome does.
}
\seealso{multinom, multiGnome}

\keyword{regression}

\examples{
library(nnet)
data("serxu")
multinom(cbind(empty, Mili, Miir, Mile, Crka) ~ habitat + night, data=serxu)$AIC
multinom(cbind(empty, Mili, Miir, Mile, Crka) ~ habitat, data=serxu)$AIC

## So there is a night effect.

## What do you think? Should automate the following?
m1 <- mergeLevels(cbind(empty, Mili, Miir, Mile, Crka)~habitat + night, serxu, "habitat")
levels(serxu$habitat)[c(3,13)] <- "cg_sb" ## Corresponds to lowest deltaAIC
m2 <- mergeLevels(cbind(empty, Mili, Miir, Mile, Crka)~habitat + night, serxu, "habitat")
levels(serxu$habitat)[c(2,10)] <- "bv_rs"
m3 <- mergeLevels(cbind(empty, Mili, Miir, Mile, Crka)~habitat + night,serxu, "habitat")
levels(serxu$habitat)[c(4,12)] <- "dg_ug"
m4 <- mergeLevels(cbind(empty, Mili, Miir, Mile, Crka)~habitat + night,serxu, "habitat")
levels(serxu$habitat)[c(5,10)] <- "fg_sx"
m5 <- mergeLevels(cbind(empty, Mili, Miir, Mile, Crka)~habitat + night,serxu, "habitat")
levels(serxu$habitat)[c(6,7)] <- "mt_mg"
m6 <- mergeLevels(cbind(empty, Mili, Miir, Mile, Crka)~habitat + night,serxu, "habitat")
levels(serxu$habitat)[c(5,8)] <- "fg_sx_pv"
m7 <- mergeLevels(cbind(empty, Mili, Miir, Mile, Crka)~habitat + night,serxu, "habitat")
levels(serxu$habitat)[c(2,8)] <- "bv_rs_sg"
m8 <- mergeLevels(cbind(empty, Mili, Miir, Mile, Crka)~habitat + night,serxu, "habitat")
min(m8)
}