\name{multiGnome-package}
\alias{multiGnome-package}
\alias{multiGnome}
\docType{package}
\title{
Indispensible for ecologists and hunters of Gnomes.
}
\description{
Fits unique regression equations to each trapped Gnome species and
returns the AIC and fitted parameters.
}
\details{
\tabular{ll}{
Package: \tab multiGnome\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2006-03-07\cr
License: \tab What license is it under?\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
David Pleydell <pleydell@supagro.inra.fr>
}
\references{
McCullagh and Nelder. Generalized Linear Models, 2nd Edition. Chapman & Hall
}


\keyword{ package }
